moved voting functions to their own file
