+++ code/devextra.h
@@ -69,10 +69,6 @@ void one_fquery (char *query);
 void plug_graveyard (CHAR_DATA *ch, int type);
 void do_pktrack (CHAR_DATA *ch, char *argument);
 bool trusts (CHAR_DATA *ch, CHAR_DATA *victim);
-void sortvotes (CHAR_DATA *ch, BALLOT_DATA *ballot);
-bool sort_votes (char *hold[], int *holdi, int cabal);
-void do_listvotes (CHAR_DATA *ch, char *argument);
-void do_vote (CHAR_DATA *ch, char *argument);
 void mob_recho (CHAR_DATA *ch, char *argument );
 void area_echo (CHAR_DATA *ch, char *echo);
 void rarea_echo (ROOM_INDEX_DATA *room, char *echo);
