+++ code/update.h
@@ -122,7 +122,6 @@ void ayell_update (void);
 /* iprog pulse update, also drift/sink code plugged in */
 void iprog_pulse_update (bool isTick);
 bool do_mob_cast (CHAR_DATA *ch);
-void save_votes (void);
 /* used for saving */
 int reverse_d (int dir);
 int get_grav_distance (OBJ_DATA *obj);
